backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.8.0
python-dotenv>=0.19.0
yfinance>=0.2.31
google-generativeai>=0.5.2
newsapi-python>=0.2.7
ta>=0.10.2
numpy==1.23.5
pandas>=2.0.3
requests>=2.31.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
torch
torchvision
torchaudio
alpaca-py>=0.8.0
scikit-learn>=1.0.2
TA-Lib>=0.4.24 